requirements.txt
google-genai>=1.0.0
httpx>=0.27.0
jinja2>=3.1.0
playwright>=1.40.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.6
